MAINT: consolidate tests in core/tests/test_shape_base
